Refactor request handlers and bot manager
